Removed unwanted code
