+++ components/SplashScreen.tsx
