--- myExpoApp/hooks/useColorScheme.ts
@@ -1,1 +1,0 @@
-export { useColorScheme } from 'react-native';
