+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "myapp",
+  "name": "myexpoapp",
   "main": "expo-router/entry",
   "version": "1.0.0",
   "scripts": {
